+++ internal/attestation/handler.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/confidential-dot-ai/c8s/internal/ear"
 	"github.com/confidential-dot-ai/c8s/pkg/attestationclient"
-	"github.com/confidential-dot-ai/c8s/pkg/operatorauth"
 	"github.com/confidential-dot-ai/c8s/pkg/ratls"
 	"github.com/confidential-dot-ai/c8s/pkg/types"
 )
@@ -24,9 +23,6 @@ type Handler struct {
 	Challenges        *ChallengeStore
 	AttestationClient attestationclient.Client
 	EarIssuer         ear.Issuer
-	// OperatorKeysHash, when set, makes /attest-key fail closed unless the
-	// caller attests the same operator-key policy. /attest remains unaffected.
-	OperatorKeysHash string
 	// RTMRs pins TDX runtime measurement registers on /attest-key: the launch
 	// digest recorded in the issued EAR covers TDVF firmware alone on TDX, so
 	// without these the EAR vouches for a guest image the host chose. Enforced
@@ -47,24 +43,13 @@ func HandleAuthenticate(challenges *ChallengeStore) http.HandlerFunc {
 
 // HandleAttestKey handles POST /attest-key: it issues an EAR (no certificate)
 // for a caller-generated ECDSA pubkey — used by in-cluster c8s components that
-// need a TEE-attested EAR for a key they generate in-process (CDS's handoff
-// signer key).
+// need a TEE-attested EAR for a key they generate in-process.
 func (h Handler) HandleAttestKey(w http.ResponseWriter, r *http.Request) {
 	var req types.AttestKeyRequestBody
 	dec := json.NewDecoder(r.Body)
 	dec.DisallowUnknownFields()
 	if err := dec.Decode(&req); err != nil {
 		WriteError(w, http.StatusUnprocessableEntity, "invalid_request", err.Error())
-		return
-	}
-	if req.OperatorKeysHash != "" {
-		if err := operatorauth.ValidateKeySetHash(req.OperatorKeysHash); err != nil {
-			WriteError(w, http.StatusBadRequest, "invalid_operator_keys_hash", err.Error())
-			return
-		}
-	}
-	if h.OperatorKeysHash != "" && req.OperatorKeysHash != h.OperatorKeysHash {
-		WriteError(w, http.StatusForbidden, "operator_policy_mismatch", "attested operator-key policy does not match this CDS")
 		return
 	}
 
@@ -94,7 +79,7 @@ func (h Handler) HandleAttestKey(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	expectedReportData, err := ratls.ReportDataForKeyWithContext(pub, challengeBytes, []byte(req.OperatorKeysHash))
+	expectedReportData, err := ratls.ReportDataForKey(pub, challengeBytes)
 	if err != nil {
 		WriteError(w, http.StatusBadRequest, "invalid_public_key", err.Error())
 		return
@@ -130,7 +115,7 @@ func (h Handler) HandleAttestKey(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	earToken, err := h.EarIssuer.IssueAttestedKey(json.RawMessage(evidenceJSON), verifyResp.Result.Claims.LaunchDigest, pub, req.OperatorKeysHash)
+	earToken, err := h.EarIssuer.IssueWithLaunchDigestAndPubKey(json.RawMessage(evidenceJSON), verifyResp.Result.Claims.LaunchDigest, pub)
 	if err != nil {
 		slog.Error("attest-key: failed to issue EAR token", "error", err)
 		WriteError(w, http.StatusInternalServerError, "ear_issuance_failed",
