+++ internal/cmds/cds/guards.go
@@ -1,20 +1,14 @@
 package cds
 
 import (
-	"github.com/confidential-dot-ai/c8s/internal/ear"
 	"github.com/confidential-dot-ai/c8s/internal/issuer"
-	"github.com/confidential-dot-ai/c8s/pkg/attestationclient"
 	"github.com/confidential-dot-ai/c8s/pkg/earsigner"
 )
 
-// Compile-time assertions that the concrete types CDS wires into the
-// internal/issuer interfaces still satisfy them. Those interfaces are defined
-// abstractly in internal/issuer to keep it decoupled from these concrete
-// packages, so the guards live here at the composition root (where both the
+// Compile-time assertion that the concrete type CDS wires into the
+// internal/issuer interface still satisfies it. The interface is defined
+// abstractly in internal/issuer to keep it decoupled from this concrete
+// package, so the guard lives here at the composition root (where both the
 // interface and the implementation are already imported) rather than next to
-// the interface definitions.
-var (
-	_ issuer.KeyProvider    = (*earsigner.Rotator)(nil)
-	_ issuer.LocalEARMinter = ear.Issuer{}
-	_ issuer.AttestationApi = attestationclient.Client{}
-)
+// the interface definition.
+var _ issuer.KeyProvider = (*earsigner.Rotator)(nil)
