+++ internal/cmds/cds/routes.go
@@ -20,7 +20,6 @@ type dependencies struct {
 	AttestKeyHandler  attestation.Handler
 	SignCSRHandler    SignCSRHandler
 	AllowlistHandler  allowlist.Handler
-	HandoffHandler    *issuer.HandoffHandler // nil disables /handoff (no --handoff-measurements)
 	ReadyFn           attestation.ReadinessFunc
 	EarIssuer         ear.Issuer
 	JWKSFunc          func() []byte
@@ -63,12 +62,6 @@ func newRouter(deps dependencies) http.Handler {
 	r.Method(http.MethodPost, "/attest", deps.protected(http.HandlerFunc(deps.AttestHandler.HandleAttest)))
 	r.Method(http.MethodPost, "/attest-key", deps.protected(http.HandlerFunc(deps.AttestKeyHandler.HandleAttestKey)))
 	r.Method(http.MethodPost, "/sign-csr", deps.protected(http.HandlerFunc(deps.SignCSRHandler.HandleSignCSR)))
-
-	// /handoff is mounted only when --handoff-measurements is set; a singleton
-	// cds runs without it.
-	if deps.HandoffHandler != nil {
-		r.Method(http.MethodPost, "/handoff", deps.protected(http.HandlerFunc(deps.HandoffHandler.HandleHandoff)))
-	}
 
 	// GET is unauthenticated (RA-TLS integrity only); every mutation goes
 	// through allowlistWrite (operator-JWT auth in the handler + rate limit +
