+++ internal/cmds/cds/run.go
@@ -13,7 +13,6 @@ import (
 	"os"
 	"os/signal"
 	"regexp"
-	"strings"
 	"syscall"
 	"time"
 
@@ -25,7 +24,6 @@ import (
 	"github.com/confidential-dot-ai/c8s/internal/readiness"
 	"github.com/confidential-dot-ai/c8s/internal/sandboxledger"
 	"github.com/confidential-dot-ai/c8s/internal/secrets"
-	pkgallowlist "github.com/confidential-dot-ai/c8s/pkg/allowlist"
 	"github.com/confidential-dot-ai/c8s/pkg/attestationclient"
 	"github.com/confidential-dot-ai/c8s/pkg/attestclient"
 	"github.com/confidential-dot-ai/c8s/pkg/certutil"
@@ -33,7 +31,6 @@ import (
 	measurementspkg "github.com/confidential-dot-ai/c8s/pkg/measurements"
 	"github.com/confidential-dot-ai/c8s/pkg/operatorauth"
 	"github.com/confidential-dot-ai/c8s/pkg/ratls"
-	"github.com/confidential-dot-ai/c8s/pkg/types"
 	"github.com/confidential-dot-ai/c8s/pkg/workloadclaims"
 	"golang.org/x/time/rate"
 )
@@ -51,8 +48,8 @@ func run(cfg config) error {
 	if err := cmdsutil.ValidateAttestationAPIURL("--attestation-api-url", cfg.attestationApiURL); err != nil {
 		return err
 	}
-	// Resolve before validateConfig: the handoff and secrets predicates read
-	// the flat lists, so a config-mode start must fill them first.
+	// Resolve before validateConfig: the secrets predicate reads the flat
+	// lists, so a config-mode start must fill them first.
 	pinned, err := resolveMeasurementsConfig(&cfg)
 	if err != nil {
 		return err
@@ -75,9 +72,6 @@ func run(cfg config) error {
 		return fmt.Errorf("init rate limiter: %w", err)
 	}
 
-	// Load the operator policy before either side of handoff is attested. Its
-	// canonical hash is committed to REPORTDATA by both replicas and compared
-	// before any CA or allowlist state is released.
 	var writeAuthorizer allowlist.WriteAuthorizer = func(*http.Request, []byte) error {
 		return fmt.Errorf("operator writes are disabled: set --operator-keys")
 	}
@@ -108,33 +102,13 @@ func run(cfg config) error {
 	}
 	defer allowlistStore.Close()
 
-	// CDS obtains its mesh CA in process; the private key never touches a
-	// Kubernetes Secret. With no --handoff-peer-url it generates a fresh
-	// self-signed CA (cold start); with a peer set it adopts that peer's CA
-	// via attested /handoff, failing closed if the peer cannot be reached or
-	// denies the handoff so a partition never mints a divergent trust root.
-	mesh, adopted, err := issuer.ProvisionCA(ctx, issuer.CAProvisionConfig{
-		CommonName:        cfg.caCommonName,
-		Validity:          cfg.caCertValidity,
-		Curve:             elliptic.P384(),
-		PeerURL:           strings.TrimRight(cfg.handoffPeerURL, "/"),
-		AttestationApiURL: cfg.attestationApiURL,
-		Measurements:      cfg.handoffMeasurements,
-		RTMRs:             cfg.handoffRTMRs,
-		ExpectedIssuer:    cfg.earIssuerName,
-		Timeout:           cfg.handoffPeerTimeout,
-		OperatorKeysHash:  operatorKeysHash,
-		RestoreAllowlist:  allowlistStore.RestoreSnapshot,
-	}, slog.Default())
+	// CDS generates its mesh CA in process at startup; the private key never
+	// touches a Kubernetes Secret.
+	mesh, err := issuer.NewCAWithCurve(cfg.caCommonName, cfg.caCertValidity, elliptic.P384())
 	if err != nil {
-		return fmt.Errorf("provision mesh CA: %w", err)
+		return fmt.Errorf("generate mesh CA: %w", err)
 	}
-	caSource := "self-generated"
-	if adopted {
-		caSource = "adopted-from-peer"
-	}
-	slog.Info("loaded in-memory mesh CA",
-		"source", caSource,
+	slog.Info("generated in-memory mesh CA",
 		"fingerprint", certutil.CertFingerprint(mesh.Cert.Raw),
 		"not_after", mesh.Cert.NotAfter.Format(time.RFC3339),
 	)
@@ -215,11 +189,7 @@ func run(cfg config) error {
 	}
 
 	if !cfg.allowlistPersistent {
-		if adopted {
-			slog.Warn("allowlist store is not durable (cds.persistence.enabled=false): this planned adoption restored the peer's dynamic entries, but a total CDS outage and deliberate re-bootstrap still resets them to the install seed")
-		} else {
-			slog.Warn("allowlist store is not persistent (cds.persistence.enabled=false): a restart without a surviving handoff peer resets the served allowlist to the install seed and regenerates the mesh CA. Operator-added digests do not survive")
-		}
+		slog.Warn("allowlist store is not persistent (cds.persistence.enabled=false): a restart resets the served allowlist to the install seed and regenerates the mesh CA. Operator-added digests do not survive")
 	}
 
 	policy := issuer.CSRPolicy{
@@ -230,15 +200,10 @@ func run(cfg config) error {
 	// /attest-key issues a TEE-attested EAR for a caller-generated key (no CSR,
 	// no certificate). Shares the challenge store, attestation-api, and EAR
 	// issuer with /attest.
-	attestKeyOperatorPolicy := ""
-	if len(cfg.handoffMeasurements) > 0 {
-		attestKeyOperatorPolicy = operatorKeysHash
-	}
 	attestKeyHandler := attestation.Handler{
 		Challenges:        &challengeStore,
 		AttestationClient: asClient,
 		EarIssuer:         earIssuer,
-		OperatorKeysHash:  attestKeyOperatorPolicy,
 		RTMRs:             rtmrPins,
 	}
 
@@ -281,11 +246,6 @@ func run(cfg config) error {
 		if err != nil {
 			return err
 		}
-	}
-
-	handoffHandler, err := buildHandoffHandler(ctx, cfg, mesh, &allowlistStore, operatorKeysHash, rotator, earIssuer, asClient)
-	if err != nil {
-		return err
 	}
 
 	// The ledger is written on every issuance, not only when secrets are on:
@@ -371,7 +331,6 @@ func run(cfg config) error {
 			MaxWriteBodyBytes: allowlistWriteBodyCap,
 		},
 		AttestKeyHandler:  attestKeyHandler,
-		HandoffHandler:    handoffHandler,
 		ReadyFn:           readinessFn(checker.Ready, mesh.Cert, cfg.minCAValidity),
 		EarIssuer:         earIssuer,
 		JWKSFunc:          rotator.JWKSetJSON,
@@ -446,74 +405,6 @@ func run(cfg config) error {
 	return nil
 }
 
-// buildHandoffHandler wires the /handoff endpoint that hands the in-memory mesh
-// CA to an attested peer replica. It is disabled (returns nil) unless
-// --handoff-measurements pins which peer launch digests may pull the CA.
-//
-// cds self-provisions its handoff signer EAR in process via
-// LocalHandoffBootstrap: cds is its own EAR issuer, so the requester EAR is
-// validated against cds's own rotator/issuer name, and the signer EAR is minted
-// by cds's earIssuer — no external service to dial for it.
-func buildHandoffHandler(ctx context.Context, cfg config, mesh *issuer.CA, allowlistStore *allowlist.Store, operatorKeysHash string, keyProvider issuer.KeyProvider, earIssuer ear.Issuer, asClient attestationclient.Client) (*issuer.HandoffHandler, error) {
-	handoffMeasurements := parseReferenceDigests(cfg.handoffMeasurements)
-	if len(handoffMeasurements) == 0 {
-		slog.Info("/handoff disabled: set --handoff-measurements to enable mesh CA handoff to peer replicas")
-		return nil, nil
-	}
-
-	boot, err := issuer.NewLocalHandoffBootstrap(asClient, earIssuer, operatorKeysHash)
-	if err != nil {
-		return nil, fmt.Errorf("prepare handoff bootstrap: %w", err)
-	}
-
-	hh, err := issuer.NewHandoffHandler(issuer.HandoffDeps{
-		Logger:              slog.Default(),
-		KeyProvider:         keyProvider,
-		ExpectedIssuer:      cfg.earIssuerName,
-		AllowedMeasurements: handoffMeasurements,
-		OperatorKeysHash:    operatorKeysHash,
-		Signer:              boot.Signer(),
-		EARSource:           boot.EARSource(),
-		Snapshot: func() (issuer.CASnapshot, bool) {
-			return snapshotAllowlist(allowlistStore.LoadAll, mesh)
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	go boot.RunRefresh(ctx, slog.Default())
-	go issuer.RunHandoffEARExpiryUpdater(ctx, hh.IssuerEARSource(), time.Minute, slog.Default())
-	slog.Info("attested CA handoff enabled (bootstrap runs in background)", "measurements", len(handoffMeasurements))
-	return hh, nil
-}
-
-// snapshotAllowlist builds the handoff CA snapshot from the live allowlist
-// store; any load or digest-parse failure withholds the snapshot.
-func snapshotAllowlist(load func() (*pkgallowlist.Allowlist, string, error), mesh *issuer.CA) (issuer.CASnapshot, bool) {
-	doc, version, err := load()
-	if err != nil {
-		slog.Error("snapshot allowlist for handoff", "error", err)
-		return issuer.CASnapshot{}, false
-	}
-	floor := make(map[types.Digest]string, len(doc.Digests))
-	for d, img := range doc.Digests {
-		pd, err := types.ParseDigest(d)
-		if err != nil {
-			slog.Error("snapshot allowlist digest parse", "digest", d, "error", err)
-			return issuer.CASnapshot{}, false
-		}
-		floor[pd] = img
-	}
-	return issuer.CASnapshot{
-		Cert:             mesh.Cert,
-		Key:              mesh.Key,
-		AllowlistVersion: version,
-		Allowlist:        floor,
-		Workloads:        doc.Workloads,
-	}, true
-}
-
 func newHTTPServer(addr string, handler http.Handler, cfg config) *http.Server {
 	cfg = normalizeHTTPServerConfig(cfg)
 	return &http.Server{
@@ -575,16 +466,8 @@ func validateSecretsConfig(cfg config) error {
 // one releases nothing and mounting the endpoint is inert until an operator
 // writes a grant. What this decides is narrower: whether CDS can answer at all,
 // which is what sandbox identity already needs.
-//
-// Handoff is the exception: a roll puts two CDS pods behind the Service at
-// once, and the surge replica serves an empty store — so a workload landing on
-// it mints a value diverging from the one its siblings already hold, with no
-// error anywhere. Refusing to serve is better than that divergence, and better
-// than failing startup, which would strand every existing handoff install.
 func secretsEnabled(cfg config, sandboxDigests *workloadclaims.DigestsClient, inventoryHosts workloadclaims.InventoryHosts) (bool, string) {
 	switch {
-	case cfg.handoffPeerURL != "" || len(cfg.handoffMeasurements) > 0:
-		return false, "handoff is configured: a surge replica would serve an empty secret store and mint values diverging from those already delivered"
 	case sandboxDigests == nil:
 		return false, "no --ratls-platform, so CDS has no attested channel to an inventory"
 	case inventoryHosts == nil || inventoryHosts.Empty():
@@ -630,22 +513,8 @@ func validateConfig(cfg config) error {
 	if cfg.readinessInterval <= 0 {
 		return fmt.Errorf("--readiness-interval must be positive")
 	}
-	if len(cfg.handoffMeasurements) > 0 && cfg.operatorKeys == "" {
-		return fmt.Errorf("--handoff-measurements requires --operator-keys so the operator policy is bound into handoff attestation")
-	}
 	if err := validateSecretsConfig(cfg); err != nil {
 		return err
-	}
-	if cfg.handoffPeerURL != "" {
-		if !strings.HasPrefix(cfg.handoffPeerURL, "https://") {
-			return fmt.Errorf("--handoff-peer-url must use https (RA-TLS)")
-		}
-		if len(cfg.handoffMeasurements) == 0 {
-			return fmt.Errorf("--handoff-peer-url requires --handoff-measurements to pin the peer")
-		}
-		if cfg.handoffPeerTimeout <= 0 {
-			return fmt.Errorf("--handoff-peer-timeout must be positive")
-		}
 	}
 	return nil
 }
