+++ internal/helmchart/c8s/templates/cds-pdb.yaml
@@ -3,10 +3,8 @@
 CDS holds the mesh CA and EAR signing key in process memory, so a voluntary
 eviction (node drain, autoscaler) of the serving pod loses the trust root.
 maxUnavailable: 0 blocks voluntary disruptions; the operator must delete the
-pod deliberately. This bounds evictions only — it does not throttle a
-RollingUpdate rollout (that obeys the Deployment's maxSurge/maxUnavailable), so
-handoff.peerUrl adoption rolls freely. Set maxUnavailable: 1 to allow drains, or
-enabled: false to remove the PDB.
+pod deliberately. Set maxUnavailable: 1 to allow drains, or enabled: false to
+remove the PDB.
 */}}
 apiVersion: policy/v1
 kind: PodDisruptionBudget
