+++ internal/issuer/log_capture_test.go
@@ -1,0 +1,60 @@
+package issuer
+
+import (
+	"context"
+	"log/slog"
+	"sync"
+)
+
+// capturedRecord is one slog record collected by captureHandler.
+type capturedRecord struct {
+	level slog.Level
+	msg   string
+	attrs map[string]slog.Value
+}
+
+// captureHandler collects records so tests can assert on structured log
+// output without string-slicing formatted lines.
+type captureHandler struct {
+	mu      sync.Mutex
+	records []capturedRecord
+}
+
+func (h *captureHandler) Enabled(context.Context, slog.Level) bool { return true }
+
+func (h *captureHandler) Handle(_ context.Context, r slog.Record) error {
+	attrs := make(map[string]slog.Value)
+	r.Attrs(func(a slog.Attr) bool {
+		attrs[a.Key] = a.Value
+		return true
+	})
+	h.mu.Lock()
+	defer h.mu.Unlock()
+	h.records = append(h.records, capturedRecord{level: r.Level, msg: r.Message, attrs: attrs})
+	return nil
+}
+
+func (h *captureHandler) WithAttrs([]slog.Attr) slog.Handler { return h }
+func (h *captureHandler) WithGroup(string) slog.Handler      { return h }
+
+func (h *captureHandler) find(msg string) (capturedRecord, bool) {
+	h.mu.Lock()
+	defer h.mu.Unlock()
+	for _, r := range h.records {
+		if r.msg == msg {
+			return r, true
+		}
+	}
+	return capturedRecord{}, false
+}
+
+func (h *captureHandler) anyAtLevel(level slog.Level) (capturedRecord, bool) {
+	h.mu.Lock()
+	defer h.mu.Unlock()
+	for _, r := range h.records {
+		if r.level >= level {
+			return r, true
+		}
+	}
+	return capturedRecord{}, false
+}
